fixing circular dependency and mypy errors
