+++ .github/workflows/mypy.yaml
@@ -21,7 +21,8 @@ jobs:
       - name: Install dependencies
         run: |
           python -m pip install --upgrade pip
-          pip install .[dev] 
+          pip install .[dev]
+          pip install mypy
 
       - name: Run mypy
         run: |
