+++ augur/tasks/github/pull_requests/tasks.py
@@ -12,7 +12,7 @@ from augur.application.db.models import PullRequest, Message, PullRequestReview,
 from augur.tasks.github.util.github_task_session import GithubTaskManifest
 from augur.tasks.github.util.github_random_key_auth import GithubRandomKeyAuth
 from augur.application.db.lib import get_session, get_repo_by_repo_git, bulk_insert_dicts, get_pull_request_reviews_by_repo_id, batch_insert_contributors
-from augur.application.db.util import execute_session_query
+from augur.application.db.helpers import execute_session_query
 from ..messages import process_github_comment_contributors
 from augur.application.db.lib import get_secondary_data_last_collected, get_updated_prs, get_core_data_last_collected
