+++ tests/test_applicaton/test_repo_load_controller/helper.py
@@ -6,7 +6,7 @@ from augur.util.repo_load_controller import ORG_REPOS_ENDPOINT
 from augur.application.db.session import DatabaseSession
 from augur.application.db.models import Config, User
 from augur.tasks.github.util.github_paginator import hit_api
-from augur.application.db.util import execute_session_query
+from augur.application.db.helpers import execute_session_query
 
 
 logger = logging.getLogger(__name__)
